Module Basic.Symbol

type t = private int

Abstract type providing an efficient encoding of some string value.

val intern : string -> t

intern s returns a symbol representing the string s.

val to_string : t -> string

to_string t returns the string associated with symbol t.