Parameter MakeExtractor.2-E

type t

Represents the type of a cost of a node.

type node

Represents terms of the language

val compare : t -> t -> int

compare c1 c2 compares the costs t1 and t2

val cost : (Id.t -> t) -> node -> t

cost f node should assign costs to the node node. It can use the provided function f to determine the cost of a child.